Refactoring views
- Move admin page views(showing html) to `views` router
- Update related routes
