+++ iap/main.py
@@ -12,7 +12,7 @@ from starlette.status import HTTP_400_BAD_REQUEST, HTTP_500_INTERNAL_SERVER_ERRO
 
 from common import logger
 from iap.exceptions import ReceiptNotFoundException
-from . import api, settings
+from . import api, views, settings
 
 __VERSION__ = "0.1.0"
 
@@ -106,25 +106,9 @@ def favicon():
     return "iap/frontend/build/favicon.png"
 
 
-@app.get("/", response_class=FileResponse, tags=["View"], summary="Index page")
-@app.get(
-    "/{page}", response_class=FileResponse, tags=["View"], summary="Opens pages provided name",
-    description="""Frontend pages are controlled by Svelte.  
-If you access to any page other than index directly from browser, this function will find right page to show.  
-
-Available page list:
-- box
-"""
-)
-def view_page(page: str = "index"):
-    # NOTICE: Set html name matches to path.
-    if os.path.isfile(f"iap/frontend/build/{page}.html"):
-        return f"iap/frontend/build/{page}.html"
-    raise HTTPException(status_code=404, detail=f"Page Not Found: /{page}")
-
-
 app.include_router(api.router)
-app.mount("/_app", StaticFiles(directory="iap/frontend/build/_app"), name="static")
+app.include_router(views.router)
+app.mount("/views/_app", StaticFiles(directory="iap/frontend/build/_app"), name="static")
 
 handler = Mangum(app)
